style: refractor the schemas
