--- server/models/data.ts
@@ -1,4 +1,0 @@
-export const users = [
-	{ id: 1, username: "alice", password: "password" },
-	{ id: 2, username: "bob", password: "123456" },
-];
